RM3100 split into separate class, implementation, and main
