+++ cmake/configs/nuttx_px4fmu-v2_default.cmake
@@ -9,7 +9,7 @@ set(config_module_list
 	#drivers/barometer
 	drivers/differential_pressure
 	#drivers/distance_sensor
-	drivers/magnetometer
+	#drivers/magnetometer
 	#drivers/telemetry
 
 	#drivers/imu/adis16448
@@ -18,7 +18,7 @@ set(config_module_list
 	#drivers/imu/bmi160
 	#drivers/barometer/bmp280
 	#drivers/bst
-	#drivers/camera_trigger
+	drivers/camera_trigger
 	#drivers/frsky_telemetry
 	drivers/gps
 	#drivers/hott
@@ -26,17 +26,17 @@ set(config_module_list
 	#drivers/irlock
 	drivers/imu/l3gd20
 	drivers/imu/lsm303d
-	#drivers/magnetometer/hmc5883
-	#drivers/magnetometer/lis3mdl
+	drivers/magnetometer/hmc5883
+	drivers/magnetometer/lis3mdl
 	#drivers/mb12xx
 	#drivers/mkblctrl
 	drivers/imu/mpu6000
-	#drivers/imu/mpu9250
+	drivers/imu/mpu9250
 	#drivers/oreoled
 	#drivers/protocol_splitter
 	drivers/pwm_input
 	#drivers/pwm_out_sim
-	#drivers/px4flow
+	drivers/px4flow
 	drivers/px4fmu
 	drivers/px4io
 	drivers/rgbled
@@ -47,13 +47,13 @@ set(config_module_list
 	drivers/vmount
 
 	# distance sensors
-	#drivers/distance_sensor/ll40ls
+	drivers/distance_sensor/ll40ls
 	#drivers/distance_sensor/mb12xx
-	#drivers/distance_sensor/sf0x
+	drivers/distance_sensor/sf0x
 	drivers/distance_sensor/sf1xx
-	#drivers/distance_sensor/srf02
-	#drivers/distance_sensor/teraranger
-	#drivers/distance_sensor/tfmini
+	drivers/distance_sensor/srf02
+	drivers/distance_sensor/teraranger
+	drivers/distance_sensor/tfmini
 	#drivers/distance_sensor/ulanding
 	modules/sensors
