+++ src/drivers/magnetometer/rm3100/CMakeLists.txt
@@ -1,6 +1,6 @@
 ############################################################################
 #
-#   Copyright (c) 2015 PX4 Development Team. All rights reserved.
+#   Copyright (c) 2018 PX4 Development Team. All rights reserved.
 #
 # Redistribution and use in source and binary forms, with or without
 # modification, are permitted provided that the following conditions
@@ -30,14 +30,13 @@
 # POSSIBILITY OF SUCH DAMAGE.
 #
 ############################################################################
+
 px4_add_module(
 	MODULE drivers__rm3100
 	MAIN rm3100
 	STACK_MAIN 1200
 	COMPILE_FLAGS
 	SRCS
-		rm3100.cpp
-	DEPENDS
-		platforms__common
+		RM3100.cpp
+		rm3100_main.cpp
 	)
-# vim: set noet ft=cmake fenc=utf-8 ff=unix :
