+++ src/drivers/magnetometer/rm3100/RM3100.cpp
@@ -1,6 +1,6 @@
 /****************************************************************************
  *
- *   Copyright (c) 2012-2016 PX4 Development Team. All rights reserved.
+ *   Copyright (c) 2018 PX4 Development Team. All rights reserved.
  *
  * Redistribution and use in source and binary forms, with or without
  * modification, are permitted provided that the following conditions
@@ -31,397 +31,20 @@
  *
  ****************************************************************************/
 
-/**
- * @file rm3100.cpp
- *
- * Driver for the RM3100 magnetometer connected via I2C.
- *
- * Author Ryan Johnston
- */
-
-#include <px4_config.h>
-#include <px4_defines.h>
-#include <px4_getopt.h>
-
-#include <sys/types.h>
-#include <stdint.h>
-#include <stdlib.h>
-#include <stdbool.h>
-#include <semaphore.h>
-#include <string.h>
-#include <fcntl.h>
-#include <poll.h>
-#include <errno.h>
-#include <stdio.h>
-#include <math.h>
-#include <unistd.h>
-
-#include <nuttx/arch.h>
-#include <nuttx/wqueue.h>
-#include <nuttx/clock.h>
-
-#include <board_config.h>
-
-#include <perf/perf_counter.h>
-#include <systemlib/err.h>
-
-#include <drivers/device/i2c.h>
-#include <drivers/drv_mag.h>
-#include <drivers/drv_hrt.h>
-#include <drivers/device/ringbuffer.h>
-#include <drivers/drv_device.h>
-
-#include <uORB/uORB.h>
-
-#include <float.h>
-#include <getopt.h>
-#include <lib/conversion/rotation.h>
-
-/*
- * RM3100 internal constants and data structures.
- */
-
-
-#define RM3100_CONVERSION_INTERVAL	(1000000 / 100)
-
-#define RM3100_BUS_I2C_ADDR		0x20
-#define RM3100_DEFAULT_BUS_SPEED	400000
-
-/*
- * FSR:
- *   x, y: +- 800 µT
- *   z:    +- 800 µT
- *
- *
- */
-#define RM3100_RESOLUTION	 0.0065
-
-static const int32_t RM3100_MAX_VAL_XY	= (800 / RM3100_RESOLUTION) + 1;
-static const int32_t RM3100_MIN_VAL_XY	= -RM3100_MAX_VAL_XY;
-static const int32_t RM3100_MAX_VAL_Z  = (800 / RM3100_RESOLUTION) + 1;
-static const int32_t RM3100_MIN_VAL_Z  = -RM3100_MAX_VAL_Z;
-
-/* Hardware definitions */
-
-//#define ADDR_WAI                0		/* WAI means 'Who Am I'*/
-//# define WAI_EXPECTED_VALUE     0xE
-
-/*#define ADDR_STAT1              0x02
-# define STAT1_DRDY_SHFITS      0x0
-# define STAT1_DRDY             (1 << STAT1_DRDY_SHFITS)
-# define STAT1_DRO_SHFITS       0x1
-# define STAT1_DRO              (1 << STAT1_DRO_SHFITS)
-*/
-
-#define ADDR_DATA_OUT_X_LSB     0x05
-#define ADDR_DATA_OUT_X_MSB     0x04
-#define ADDR_DATA_OUT_Y_LSB     0x07
-#define ADDR_DATA_OUT_Y_MSB     0x06
-#define ADDR_DATA_OUT_Z_LSB     0x09
-#define ADDR_DATA_OUT_Z_MSB     0x08
-
-/*#define ADDR_STAT2              0x09
-# define STAT2_INT_SHFITS       3
-# define STAT2_INT              (1 << STAT2_INT_SHFITS)
-*/
-
-# define ADDR_CTRL1              0x00
-//# define CTRL1_MODE_SHFITS      0
-//# define CTRL1_MODE_STDBY       (0 << CTRL1_MODE_SHFITS)
-# define CTRL1_MODE_SINGLE      0x70
-
-# define ADDR_CTRL2              0x34
-//# define CTRL2_SRST_SHFITS      0   /* Begin POR (auto cleared) */
-//# define CTRL2_SRST             (1 << CTRL2_SRST_SHFITS)
-//# define CTRL2_DRP_SHIFTS       2
-//# define CTRL2_DRP              (1 << CTRL2_DRP_SHIFTS)
-//# define CTRL2_DREN_SHIFTS      3
-//# define CTRL2_DREN             (1 << CTRL2_DREN_SHIFTS)
-
-#define ADDR_CTRL3				0x0B
-# define CTRL3_SAMPLEAVG_16		0x96	/* Sample Averaging 16 */
-//# define CTRL3_SAMPLEAVG_8		0x1b	/* Sample Averaging 8 */
-//# define CTRL3_SAMPLEAVG_4		0x12	/* Sample Averaging 4 */
-//# define CTRL3_SAMPLEAVG_2		0x09	/* Sample Averaging 2 */
-
-//#define ADDR_CTRL4				0x42
-//# define CTRL4_SRPD				0xC0	/* Set Reset Pulse Duration */
-
-#define ADDR_STR                0x33
-# define STR_SELF_TEST_SHFITS   8
-# define STR_SELF_TEST_ON       (1 << STR_SELF_TEST_SHFITS)
-//# define STR_SELF_TEST_OFF      (0 << STR_SELF_TEST_SHFITS)
-
-/*#define ADDR_Y11_Low			0x9c
-#define ADDR_Y11_High			0x9d
-#define ADDR_Y12_Low			0x9e
-#define ADDR_Y12_High			0x9f
-#define ADDR_Y13_Low			0xa0
-#define ADDR_Y13_High			0xa1
-#define ADDR_Y21_Low			0xa2
-#define ADDR_Y21_High			0xa3
-#define ADDR_Y22_Low			0xa4
-#define ADDR_Y22_High			0xa5
-#define ADDR_Y23_Low			0xa6
-#define ADDR_Y23_High			0xa7
-#define ADDR_Y31_Low			0xa8
-#define ADDR_Y31_High			0xa9
-#define ADDR_Y32_Low			0xaa
-#define ADDR_Y32_High			0xab
-#define ADDR_Y33_Low			0xac
-#define ADDR_Y33_High			0xad
-*/
-
-//#define ADDR_TEMPL              0x1c
-//#define ADDR_TEMPH              0x1d
-
-enum RM3100_BUS {
-	RM3100_BUS_ALL           = 0,
-	RM3100_BUS_I2C_EXTERNAL  = 1,
-	RM3100_BUS_I2C_EXTERNAL1 = 2,
-	RM3100_BUS_I2C_EXTERNAL2 = 3,
-	RM3100_BUS_I2C_EXTERNAL3 = 4,
-	RM3100_BUS_I2C_INTERNAL  = 5,
-};
-
-#ifndef CONFIG_SCHED_WORKQUEUE
-# error This requires CONFIG_SCHED_WORKQUEUE.
-#endif
-
-class RM3100 : public device::I2C
-{
-public:
-	RM3100(int bus_number, int address, const char *path, enum Rotation rotation);
-	virtual ~RM3100();
-
-	virtual int     init();
-
-	virtual ssize_t     read(struct file *filp, char *buffer, size_t buflen);
-	virtual int         ioctl(struct file *filp, int cmd, unsigned long arg);
-
-	/**
-	 * Diagnostics - print some basic information about the driver.
-	 */
-	void            print_info();
-
-//protected:
-//	virtual int probe();
-
-private:
-	work_s          _work;
-	unsigned        _measure_ticks;
-
-	ringbuffer::RingBuffer  *_reports;
-	struct mag_calibration_s    _scale;
-	float           _range_scale;
-	bool        _collect_phase;
-	int         _class_instance;
-	int         _orb_class_instance;
-
-	orb_advert_t        _mag_topic;
-
-	perf_counter_t      _sample_perf;
-	perf_counter_t      _comms_errors;
-	perf_counter_t      _range_errors;
-	perf_counter_t      _conf_errors;
-
-	/* status reporting */
-	bool            _sensor_ok;         /**< sensor was found and reports ok */
-	bool            _calibrated;        /**< the calibration is valid */
-	bool			_ctl_reg_mismatch;	/**< control register value mismatch after checking */
-
-	enum Rotation       _rotation;
-
-	struct mag_report   _last_report;           /**< used for info() */
-
-	uint8_t 		_ctl3_reg;
-	uint8_t			_ctl4_reg;
-
-	/**
-	 * Initialise the automatic measurement state machine and start it.
-	 *
-	 * @note This function is called at open and error time.  It might make sense
-	 *       to make it more aggressive about resetting the bus in case of errors.
-	 */
-	void		start();
-
-	/**
-	 * Stop the automatic measurement state machine.
-	 */
-	void        stop();
-
-	/**
-	 * Reset the device
-	 */
-	int         reset();
-
-	/**
-	 * Perform the on-sensor scale calibration routine.
-	 *
-	 * @note The sensor will continue to provide measurements, these
-	 *   will however reflect the uncalibrated sensor state until
-	 *   the calibration routine has been completed.
-	 *
-	 * @param enable set to 1 to enable self-test strap, 0 to disable
-	 */
-	int         calibrate(struct file *filp, unsigned enable);
-
-	/**
-	 * check the sensor configuration.
-	 *
-	 * checks that the config of the sensor is correctly set, to
-	 * cope with communication errors causing the configuration to
-	 * change
-	 */
-	void            check_conf(void);
-
-	/**
-	 * Perform a poll cycle; collect from the previous measurement
-	 * and start a new one.
-	 *
-	 * This is the heart of the measurement state machine.  This function
-	 * alternately starts a measurement, or collects the data from the
-	 * previous measurement.
-	 *
-	 * When the interval between measurements is greater than the minimum
-	 * measurement interval, a gap is inserted between collection
-	 * and measurement to provide the most recent measurement possible
-	 * at the next interval.
-	 */
-	void            cycle();
-
-	/**
-	 * Static trampoline from the workq context; because we don't have a
-	 * generic workq wrapper yet.
-	 *
-	 * @param arg       Instance pointer for the driver that is polling.
-	 */
-	static void     cycle_trampoline(void *arg);
-
-	/**
-	 * Write a register.
-	 *
-	 * @param reg       The register to write.
-	 * @param val       The value to write.
-	 * @return      OK on write success.
-	 */
-	int         write_reg(uint8_t reg, uint8_t val);
-
-	/**
-	 * Write to a register block.
-	 *
-	 * @param address   The register address to write to.
-	 * @param data      The buffer to write from.
-	 * @param count     The number of bytes to write.
-	 * @return      OK on write success.
-	 */
-	int     write(unsigned address, void *data, unsigned count);
-
-	/**
-	 * Read a register.
-	 *
-	 * @param reg       The register to read.
-	 * @param val       The value read.
-	 * @return      OK on read success.
-	 */
-	int         read_reg(uint8_t reg, uint8_t &val);
-
-	/**
-	 * read register block.
-	 *
-	 * @param address   The register address to read from.
-	 * @param data      The buffer to read into.
-	 * @param count     The number of bytes to read.
-	 * @return      OK on write success.
-	 */
-	int read(unsigned address, void *data, unsigned count);
-
-	/**
-	 * Issue a measurement command.
-	 *
-	 * @return      OK if the measurement command was successful.
-	 */
-	int         measure();
-
-	/**
-	 * Collect the result of the most recent measurement.
-	 */
-	int         collect();
-
-	/**
-	 * Convert a big-endian signed 16-bit value to a float.
-	 *
-	 * @param in        A signed 16-bit big-endian value.
-	 * @return      The floating-point representation of the value.
-	 */
-	float       meas_to_float(uint8_t in[2]);
-
-	/**
-	 * Check the current calibration and update device status
-	 *
-	 * @return 0 if calibration is ok, 1 else
-	 */
-	int         check_calibration();
-
-	/**
-	* Check the current scale calibration
-	*
-	* @return 0 if scale calibration is ok, 1 else
-	*/
-	int         check_scale();
-
-	/**
-	* Check the current offset calibration
-	*
-	* @return 0 if offset calibration is ok, 1 else
-	*/
-	int         check_offset();
-
-	/**
-	* Place the device in self test mode
-	*
-	* @return 0 if mode is set, 1 else
-	*/
-	int         set_selftest(unsigned enable);
-
-	/* this class has pointer data members, do not allow copying it */
-	RM3100(const RM3100 &);
-	RM3100 operator=(const RM3100 &);
-};
-
-/*
- * Driver 'main' command.
- */
-extern "C" __EXPORT int rm3100_main(int argc, char *argv[]);
-
+#include "RM3100.hpp"
 
 RM3100::RM3100(int bus_number, int address, const char *path, enum Rotation rotation) :
 	I2C("RM3100", path, bus_number, address, RM3100_DEFAULT_BUS_SPEED),
-	_work{},
-	_measure_ticks(0),
-	_reports(nullptr),
-	_scale{},
-	_range_scale(0.003), /* default range scale from counts to gauss */
-	_collect_phase(false),
-	_class_instance(-1),
-	_orb_class_instance(-1),
-	_mag_topic(nullptr),
 	_sample_perf(perf_alloc(PC_ELAPSED, "rm3100_read")),
 	_comms_errors(perf_alloc(PC_COUNT, "rm3100_com_err")),
 	_range_errors(perf_alloc(PC_COUNT, "rm3100_rng_err")),
 	_conf_errors(perf_alloc(PC_COUNT, "rm3100_conf_err")),
-	_sensor_ok(false),
-	_calibrated(false),
-	_ctl_reg_mismatch(false),
-	_rotation(rotation),
-	_last_report{0},
-	_ctl3_reg(0),
-	_ctl4_reg(0)
+	_rotation(rotation)
 {
 	_device_id.devid_s.devtype = DRV_MAG_DEVTYPE_RM3100;
 
 	// enable debug() calls
-	_debug_enabled = false;
+	_debug_enabled = true;
 
 	// default scaling
 	_scale.x_offset = 0;
@@ -430,9 +53,6 @@ RM3100::RM3100(int bus_number, int address, const char *path, enum Rotation rota
 	_scale.y_scale = 1.0f;
 	_scale.z_offset = 0;
 	_scale.z_scale = 1.0f;
-
-	// work_cancel in the dtor will explode if we don't do this...
-	memset(&_work, 0, sizeof(_work));
 }
 
 RM3100::~RM3100()
@@ -537,7 +157,7 @@ void RM3100::check_conf(void)
 		_ctl_reg_mismatch = true;
 	}
 
-	
+
 	_ctl_reg_mismatch = false;
 }
 
@@ -688,7 +308,7 @@ RM3100::ioctl(struct file *filp, int cmd, unsigned long arg)
 		}
 
 	//case SENSORIOCRESET:
-		//return reset();
+	//return reset();
 
 	case MAGIOCSSAMPLERATE:
 		/* same as pollrate because device is in single measurement mode*/
@@ -756,14 +376,14 @@ RM3100::stop()
 //int
 //RM3100::reset()
 //{
-	/* software reset */
+/* software reset */
 //	write_reg(ADDR_CTRL2);
 
-	/* configure control register 3 */
+/* configure control register 3 */
 //	_ctl3_reg = CTRL3_SAMPLEAVG_16;
 //	write_reg(ADDR_CTRL3, _ctl3_reg);
 
-	/* configure control register 4 */
+/* configure control register 4 */
 //	_ctl4_reg = CTRL4_SRPD;
 //	write_reg(ADDR_CTRL4, _ctl4_reg);
 
@@ -1243,396 +863,4 @@ RM3100::print_info()
 	printf("poll interval:  %u ticks\n", _measure_ticks);
 	print_message(_last_report);
 	_reports->print_info("report queue");
-}
-
-/**
- * Local functions in support of the shell command.
- */
-namespace rm3100
-{
-
-/*
-  list of supported bus configurations
- */
-struct rm3100_bus_option {
-	enum RM3100_BUS busid;
-	const char *devpath;
-	uint8_t busnum;
-	RM3100 *dev;
-} bus_options[] = {
-	{ RM3100_BUS_I2C_EXTERNAL, "/dev/rm3100_ext", PX4_I2C_BUS_EXPANSION, NULL },
-#ifdef PX4_I2C_BUS_EXPANSION1
-	{ RM3100_BUS_I2C_EXTERNAL1, "/dev/ist8311_ext1", PX4_I2C_BUS_EXPANSION1, NULL },
-#endif
-#ifdef PX4_I2C_BUS_EXPANSION2
-	{ RM3100_BUS_I2C_EXTERNAL2, "/dev/ist8312_ext2", PX4_I2C_BUS_EXPANSION2, NULL },
-#endif
-#ifdef PX4_I2C_BUS_ONBOARD
-	{ RM3100_BUS_I2C_INTERNAL, "/dev/rm3100_int", PX4_I2C_BUS_ONBOARD, NULL },
-#endif
-};
-#define NUM_BUS_OPTIONS (sizeof(bus_options)/sizeof(bus_options[0]))
-
-void    start(enum RM3100_BUS busid, int address, enum Rotation rotation);
-bool    start_bus(struct rm3100_bus_option &bus, int address, enum Rotation rotation);
-struct rm3100_bus_option &find_bus(enum RM3100_BUS busid);
-void    test(enum RM3100_BUS busid);
-void    reset(enum RM3100_BUS busid);
-int info(enum RM3100_BUS busid);
-int calibrate(enum RM3100_BUS busid);
-void    usage();
-
-/**
- * start driver for a specific bus option
- */
-bool
-start_bus(struct rm3100_bus_option &bus, int address, enum Rotation rotation)
-{
-	if (bus.dev != nullptr) {
-		errx(1, "bus option already started");
-	}
-
-	RM3100 *interface = new RM3100(bus.busnum, address,  bus.devpath, rotation);
-
-	if (interface->init() != OK) {
-		delete interface;
-		PX4_INFO("no device on bus %u", (unsigned)bus.busid);
-		return false;
-	}
-
-	bus.dev = interface;
-
-	int fd = open(bus.devpath, O_RDONLY);
-
-	if (fd < 0) {
-		return false;
-	}
-
-	if (ioctl(fd, SENSORIOCSPOLLRATE, SENSOR_POLLRATE_DEFAULT) < 0) {
-		close(fd);
-		errx(1, "Failed to setup poll rate");
-	}
-
-	close(fd);
-
-	return true;
-}
-
-
-/**
- * Start the driver.
- *
- * This function call only returns once the driver
- * is either successfully up and running or failed to start.
- */
-void
-start(enum RM3100_BUS busid, int address, enum Rotation rotation)
-{
-	bool started = false;
-
-	for (unsigned i = 0; i < NUM_BUS_OPTIONS; i++) {
-		if (busid == RM3100_BUS_ALL && bus_options[i].dev != NULL) {
-			// this device is already started
-			continue;
-		}
-
-		if (busid != RM3100_BUS_ALL && bus_options[i].busid != busid) {
-			// not the one that is asked for
-			continue;
-		}
-
-		started |= start_bus(bus_options[i], address, rotation);
-	}
-
-	if (!started) {
-		exit(1);
-	}
-}
-
-/**
- * find a bus structure for a busid
- */
-struct rm3100_bus_option &find_bus(enum RM3100_BUS busid)
-{
-	for (unsigned i = 0; i < NUM_BUS_OPTIONS; i++) {
-		if ((busid == RM3100_BUS_ALL ||
-		     busid == bus_options[i].busid) && bus_options[i].dev != NULL) {
-			return bus_options[i];
-		}
-	}
-
-	errx(1, "bus %u not started", (unsigned)busid);
-}
-
-
-/**
- * Perform some basic functional tests on the driver;
- * make sure we can collect data from the sensor in polled
- * and automatic modes.
- */
-void
-test(enum RM3100_BUS busid)
-{
-	struct rm3100_bus_option &bus = find_bus(busid);
-	struct mag_report report;
-	ssize_t sz;
-	int ret;
-	const char *path = bus.devpath;
-
-	int fd = open(path, O_RDONLY);
-
-	if (fd < 0) {
-		err(1, "%s open failed (try 'rm3100 start')", path);
-	}
-
-	/* do a simple demand read */
-	sz = read(fd, &report, sizeof(report));
-
-	if (sz != sizeof(report)) {
-		err(1, "immediate read failed");
-	}
-
-	print_message(report);
-
-	/* check if mag is onboard or external */
-	if ((ret = ioctl(fd, MAGIOCGEXTERNAL, 0)) < 0) {
-		errx(1, "failed to get if mag is onboard or external");
-	}
-
-	/* set the queue depth to 5 */
-	if (OK != ioctl(fd, SENSORIOCSQUEUEDEPTH, 10)) {
-		errx(1, "failed to set queue depth");
-	}
-
-	/* start the sensor polling at 2Hz */
-	if (OK != ioctl(fd, SENSORIOCSPOLLRATE, 2)) {
-		errx(1, "failed to set 2Hz poll rate");
-	}
-
-	/* read the sensor 5x and report each value */
-	for (unsigned i = 0; i < 5; i++) {
-		struct pollfd fds;
-
-		/* wait for data to be ready */
-		fds.fd = fd;
-		fds.events = POLLIN;
-		ret = poll(&fds, 1, 2000);
-
-		if (ret != 1) {
-			errx(1, "timed out waiting for sensor data");
-		}
-
-		/* now go get it */
-		sz = read(fd, &report, sizeof(report));
-
-		if (sz != sizeof(report)) {
-			err(1, "periodic read failed");
-		}
-
-		print_message(report);
-	}
-
-	PX4_INFO("PASS");
-	exit(0);
-}
-
-
-/**
- * Automatic scale calibration.
- *
- * Basic idea:
- *
- *   output = (ext field +- 1.1 Ga self-test) * scale factor
- *
- * and consequently:
- *
- *   1.1 Ga = (excited - normal) * scale factor
- *   scale factor = (excited - normal) / 1.1 Ga
- *
- *   sxy = (excited - normal) / 766 | for conf reg. B set to 0x60 / Gain = 3
- *   sz  = (excited - normal) / 713 | for conf reg. B set to 0x60 / Gain = 3
- *
- * By subtracting the non-excited measurement the pure 1.1 Ga reading
- * can be extracted and the sensitivity of all axes can be matched.
- *
- * SELF TEST OPERATION
- * To check the RM3100L for proper operation, a self test feature in incorporated
- * in which the sensor will change the polarity on all 3 axis. The values with and
- * with and without selftest on shoult be compared and if the absolete value are equal
- * the IC is functional.
- */
-int calibrate(enum RM3100_BUS busid)
-{
-	int ret;
-	struct rm3100_bus_option &bus = find_bus(busid);
-	const char *path = bus.devpath;
-
-	int fd = open(path, O_RDONLY);
-
-	if (fd < 0) {
-		err(1, "%s open failed (try 'rm3100 start' if the driver is not running", path);
-	}
-
-	if (OK != (ret = ioctl(fd, MAGIOCCALIBRATE, fd))) {
-		PX4_WARN("failed to enable sensor calibration mode");
-	}
-
-	close(fd);
-
-	return ret;
-}
-
-/**
- * Reset the driver.
- */
-void
-reset(enum RM3100_BUS busid)
-{
-	struct rm3100_bus_option &bus = find_bus(busid);
-	const char *path = bus.devpath;
-
-	int fd = open(path, O_RDONLY);
-
-	if (fd < 0) {
-		err(1, "failed ");
-	}
-
-	if (ioctl(fd, SENSORIOCRESET, 0) < 0) {
-		err(1, "driver reset failed");
-	}
-
-	if (ioctl(fd, SENSORIOCSPOLLRATE, SENSOR_POLLRATE_DEFAULT) < 0) {
-		err(1, "driver poll restart failed");
-	}
-
-	// Relay on at_)exit to close handel
-	exit(0);
-}
-
-
-
-/**
- * Print a little info about the driver.
- */
-int
-info(enum RM3100_BUS busid)
-{
-	struct rm3100_bus_option &bus = find_bus(busid);
-
-	PX4_INFO("running on bus: %u (%s)\n", (unsigned)bus.busid, bus.devpath);
-	bus.dev->print_info();
-	exit(0);
-}
-
-void
-usage()
-{
-	PX4_INFO("missing command: try 'start', 'info', 'test', 'reset', 'calibrate'");
-	PX4_INFO("options:");
-	PX4_INFO("    -R rotation");
-	PX4_INFO("    -C calibrate on start");
-	PX4_INFO("    -a 12C Address (0x%02x)", RM3100_BUS_I2C_ADDR);
-	PX4_INFO("    -b 12C bus (%d-%d)", RM3100_BUS_I2C_EXTERNAL, RM3100_BUS_I2C_INTERNAL);
-}
-
-} // namespace
-
-int
-rm3100_main(int argc, char *argv[])
-{
-	RM3100_BUS i2c_busid = RM3100_BUS_ALL;
-	int i2c_addr = RM3100_BUS_I2C_ADDR; /* 7bit */
-
-	enum Rotation rotation = ROTATION_NONE;
-	bool calibrate = false;
-	int myoptind = 1;
-	int ch;
-	const char *myoptarg = nullptr;
-
-	while ((ch = px4_getopt(argc, argv, "R:Ca:b:", &myoptind, &myoptarg)) != EOF) {
-		switch (ch) {
-		case 'R':
-			rotation = (enum Rotation)atoi(myoptarg);
-			break;
-
-		case 'a':
-			i2c_addr = (int)strtol(myoptarg, NULL, 0);
-			break;
-
-		case 'b':
-			i2c_busid = (RM3100_BUS)strtol(myoptarg, NULL, 0);
-			break;
-
-		case 'C':
-			calibrate = true;
-			break;
-
-		default:
-			rm3100::usage();
-			exit(0);
-		}
-	}
-
-	if (myoptind >= argc) {
-		rm3100::usage();
-		return 1;
-	}
-
-	const char *verb = argv[myoptind];
-
-	/*
-	 * Start/load the driver.
-	 */
-	if (!strcmp(verb, "start")) {
-		rm3100::start(i2c_busid, i2c_addr, rotation);
-
-		if (i2c_busid == RM3100_BUS_ALL) {
-			PX4_ERR("calibration only feasible against one bus");
-			return 1;
-
-		} else if (calibrate && (rm3100::calibrate(i2c_busid) != 0)) {
-			PX4_ERR("calibration failed");
-			return 1;
-		}
-
-		return 0;
-	}
-
-	/*
-	 * Test the driver/device.
-	 */
-	if (!strcmp(verb, "test")) {
-		rm3100::test(i2c_busid);
-	}
-
-	/*
-	 * Reset the driver.
-	 */
-	if (!strcmp(verb, "reset")) {
-		rm3100::reset(i2c_busid);
-	}
-
-	/*
-	 * Print driver information.
-	 */
-	if (!strcmp(verb, "info") || !strcmp(verb, "status")) {
-		rm3100::info(i2c_busid);
-	}
-
-	/*
-	 * Autocalibrate the scaling
-	 */
-	if (!strcmp(verb, "calibrate")) {
-		if (rm3100::calibrate(i2c_busid) == 0) {
-			PX4_INFO("calibration successful");
-			return 0;
-
-		} else {
-			PX4_ERR("calibration failed");
-			return 1;
-		}
-	}
-
-	rm3100::usage();
-	return 1;
 }
